+++ src/Zicht/Itertools/lib/Traits/SliceTrait.php
@@ -12,8 +12,8 @@ trait SliceTrait
     /**
      * TODO: document!
      *
-     * @param integer $start
-     * @param null|integer $end
+     * @param int $start
+     * @param null|int $end
      * @return SliceIterator
      */
     public function slice($start, $end = null)