+++ tests/Zicht/ItertoolsTest/FirstTest.php
@@ -115,9 +115,13 @@ class FirstTest extends \PHPUnit_Framework_TestCase
             [[0]],
             [[1.0]],
             [[true]],
-            [[function () {
-                return [];
-            }]],
+            [
+                [
+                    function () {
+                        return [];
+                    },
+                ],
+            ],
         ];
     }
 }